Refactor: #4 Ask vuefile move to admin from franchise and refactor code
파일 위치 변경 및 css 리펙토링
